refactor(frontend): share quiz summary sidebar component
