+++ assets/src/js/frontend/components/quiz/summary-sidebar.ts
