+++ assets/src/js/frontend/dashboard/pages/quiz-attempts.ts
@@ -1,5 +1,6 @@
 // Quiz Attempts Page
 import { type MutationState } from '@Core/ts/services/Query';
+import { quizSummarySidebarMeta } from '@FrontendComponents/quiz/summary-sidebar';
 import { tutorConfig } from '@TutorShared/config/config';
 import { wpAjaxInstance } from '@TutorShared/utils/api';
 import { convertToErrorMessage } from '@TutorShared/utils/util';
@@ -60,12 +61,14 @@ const quizAttemptsPage = () => {
 };
 
 export const initializeQuizAttempts = () => {
-  window.TutorComponentRegistry.register({
-    type: 'component',
-    meta: {
-      name: 'quizAttempts',
-      component: quizAttemptsPage,
-    },
+  window.TutorComponentRegistry.registerAll({
+    components: [
+      {
+        name: 'quizAttempts',
+        component: quizAttemptsPage,
+      },
+      quizSummarySidebarMeta,
+    ],
   });
 
   window.TutorComponentRegistry.initWithAlpine(window.Alpine);
